MacMerch/requirements.txt
pandas
requests
dotenv
python-dotenv
google-cloud-bigquery
google-auth
google-auth-oauthlib